fix wording: asset market -> assets
